Switch from int to float size

+++ app/src/main/java/com/activityartapp/presentation/common/ScreenMeasurer.kt
@@ -1,13 +1,11 @@
 package com.activityartapp.presentation.common
 
-import android.util.Size
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.SideEffect
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.platform.LocalDensity
 
 @Composable
@@ -16,7 +14,7 @@ fun ScreenMeasurer(onMeasured: (Size) -> Unit) {
         val localDensity = LocalDensity.current
         SideEffect {
             onMeasured(localDensity.run {
-                Size(maxWidth.toPx().toInt(), maxHeight.toPx().toInt())
+                Size(maxWidth.toPx(), maxHeight.toPx())
             })
         }
     }
